mp: zaud: fix 16-bit gain overflow above unity
apply_gain_16bit() computed the gain product in a 32-bit intermediate via
GAIN_MULTIPLY(). gain_fixed is Q16.16, so for a gain above 100% and
|sample| > 16384 the product exceeds INT32_MAX and wraps before the clamp
can saturate it. The worst case is silent: at 200% gain a full-scale
INT16_MIN sample yields 0 instead of INT16_MIN.

Sweeping the full int16 range shows how much of the output is affected:
150% gain corrupts 33% of samples, 200% corrupts 50% and 1000% corrupts
90%. Gains at or below unity were never affected, which is why the
sample's default of 90% hid this.

Promote to a 64-bit intermediate, matching what apply_gain_32bit() and
apply_gain_24bit() already do.

+++ subsys/mp/plugins/zaud/mp_zaud_gain.c
@@ -24,9 +24,11 @@ LOG_MODULE_REGISTER(mp_zaud_gain, CONFIG_MP_LOG_LEVEL);
 #define GAIN_PERCENT_MAX   1000 /* 1000% = 10x amplification */
 #define GAIN_PERCENT_UNITY 100  /* 100% = unity gain (no change) */
 
-/* Q16.16 signed fixed-point definitions */
+/* Q16.16 signed fixed-point. GAIN_MULTIPLY promotes to 64 bits: a full-scale
+ * sample times a gain above unity overflows a 32-bit product.
+ */
 #define GAIN_UNITY_FIXED            (1 << 16) /* 65536 = 1.0 in Q16.16 */
-#define GAIN_MULTIPLY(sample, gain) (((int32_t)(sample) * (gain)) >> 16)
+#define GAIN_MULTIPLY(sample, gain) (((int64_t)(sample) * (gain)) >> 16)
 
 /* Convert percentage to Q16.16 fixed-point gain */
 static int32_t percent_to_fixed_gain(int gain_percent)
@@ -86,7 +88,7 @@ static void apply_gain_16bit(struct net_buf *buffer, int32_t gain_fixed)
 
 	for (size_t i = 0; i < num_samples; i++) {
 		/* Apply gain using fixed-point arithmetic */
-		int32_t temp = GAIN_MULTIPLY(samples[i], gain_fixed);
+		int64_t temp = GAIN_MULTIPLY(samples[i], gain_fixed);
 
 		/* Clamp to 16-bit range */
 		if (temp > INT16_MAX) {
